Completely redesigned cars array using linked list.
Linked list implementation supports push and pop from each end. Push/pop in the middle are prototyped but not yet implemented

Removed neighbour list implementation as not compatible. A new implementation should be designed to speed up the simulation.

Removed the sorting functions as not compatible (and not required) with the linked list implementation

Removed temporary lines, some debug code, cleaned up the code.